vosk-transcriber: fix JSON output
The variable name was incorrect. This led to an error when the output
was written. Also the result of appending a string to an array is that
each character of the string becomes a separate element in the array.
Appending the string as a single array element is much more useful.